code refactored

+++ DAL/Interfaces/Service/IProductService.cs
@@ -1,5 +1,4 @@
 ﻿using BLL.DTO;
-using Ecommerce.Models;
 using Ecommerce.Models.DTO;
 
 namespace Ecommerce.Repositories
@@ -7,9 +6,9 @@ namespace Ecommerce.Repositories
     public interface IProductService
     {
         Task<List<ProductDTO>> GetAllProducts();
-        Task<ProductDTO> GetProductById(long productId);
+        Task<ProductDTO> GetProductById(int productId);
         Task<ProductDTO> CreateProduct(CreateProductDTO createProductDTO);
-        Task<ProductDTO> UpdateProduct(CreateProductDTO updatedProduct, long id);
-        Task<bool> DeleteProductById(long productId);
+        Task<ProductDTO> UpdateProduct(CreateProductDTO updatedProduct, int id);
+        Task<bool> DeleteProductById(int productId);
     }
 }
